feat(ui): make the two destinations pages, and let the dock read the swipe
Home and Accounts were two states of one slot, reachable only by aiming at a
dock item. They are pages now: the finger moves between them directly, which is
the shortest path between the only two places the shell has.

The dock takes a position rather than an index. Colour, label weight, the
outline→filled glyph and one thin ledger rule all interpolate on the pager's
scroll fraction, so the bar answers the gesture while it is happening instead of
flipping at the end of it. The rule grows on the destination being reached as it
drains from the one being left — one mark travelling, not two blinking.

Back on Accounts drags that pager instead of shrinking the whole app: returning
to a peer is not leaving a screen, and the hierarchy pull would say it was.
Committing moves the shell's own tab, which is the path a dock tap takes too, so
both ways of going back settle identically.

The neighbouring page stays composed so a swipe starts on a drawn screen, but
while it is off-screen it is hidden from accessibility — otherwise the semantics
tree holds two Accounts screens, one of them invisible, and the UIAutomator
journey clicked the invisible one.

+++ app/src/test/java/dev/whekin/whfin/ui/LedgerDockTest.kt
@@ -30,7 +30,8 @@ class LedgerDockTest {
         var addRequests = 0
         compose.setContent {
             WhfinTheme {
-                LedgerDock(selected, onAdd = { addRequests += 1 }) { selected = it }
+                // The dock reads a page position, so a settled tab is that page's whole number.
+                LedgerDock(selected.toFloat(), onAdd = { addRequests += 1 }) { selected = it }
             }
         }
